refactor(storage): share source-derived run projection reads
Problem: run, observed-event, and context-snapshot reads had two query substrates. ArchiveStore terminal queries synthesized source rows from sessions/blocks, while repository/query-store reads selected only the materialized session_* cache tables. The duplicated SQL also meant absent cache tables failed even when source evidence was enough to answer the query.

What changed: add a shared run_projection_relations module with source-derived relation builders, optional materialized branches, and row mappers. Route ArchiveStore terminal queries, structural exists selectors, and async query-store list APIs through that shared relation. Keep materialized subagent/enrichment rows when the cache tables exist, while direct repository reads still return source rows when the tables are empty or absent.

Verification: ruff format/check on touched files; python -m py_compile for archive/query/shared modules; mypy on the shared and async query modules; devtools test tests/unit/insights/test_run_projection_materialization.py tests/unit/cli/test_query_expression.py::TestBooleanQueryExpression::test_exists_run_projection_predicates_select_sessions tests/unit/cli/test_query_expression.py::TestBooleanQueryExpression::test_exists_run_projection_predicates_use_source_relations -> 6 passed in 76.2s.
